moved functionality from controller to service
